rfe9249, bug18948 - Fix Python namespace implementation and part parsing
RELNOTES was added and updated to explain user visible changes

All nosetests pass, but test22 was changed to SKIP since it can hang
in the server. It will be restored when bug18964 is fixed.

The namespace interface was changed to use the server-side calls instead
of a client-side map. This is more consistent now with other AG40 client
implementations.

Part parsing of ntriples is now cleaner and more complete than before.

The JDBCResultSet interface was also removed in this commit as it is
an unnecessary optimization. Tests and tutorials were updated.