Added auto-commit-on-size interface to Python client - rfe9295
Updated the miniclient to include the commit parameter on the
appropriate statements requests. At the openrdf level, there
is a new property on RepositoryConnection that lets you set
the add_commit_size, which is then used by addStatements and
addFile. It was done this way in case the feature is made
obsolete someday we wouldn't have to reduce the number of
parameters on the add functions.

Tests added for:     rfe9295 - test_add_commit_size in tests.py
make test-suite run? yes

<release-note>
rfe9295: Changes addStatements and addFile

addStatements and addFile can now auto-commit triples based on a
count for large loads of triples. Set the add commit size to a positive
integer by calling setAddCommitSize or setting the add_commit_size
property on the RepositoryConnection for size-based commits, or
set to None or 0 for regular commit semantic on the add methods.

Using this feature can decrease the server's memory usage during loading.
</release-note>

Signed-off-by: Ahmon Dancy <dancy@franz.com>