rfe9318: Move to session-based federation
This updates the client, its test-suites, and the tutorial to use
sessions for federation, rather than creating a persistent federated
store and treating it like an on-disk store.

<release-note>
rfe9318: Move to session-based federation API.

To federate stores, one now calls AllegroGraphServer.openFederated,
which opens a session that federates the given stores, rather than
creating a persistent federated store and opening that.
</release-note>

<documentation>
I've opened rfe9378 for Bruce to update the docs.
</documentation>

Signed-off-by: Ahmon Dancy <dancy@franz.com>