Python tests used fixed file names in /tmp - bug19110
tests.py:test8 was changed to write to stdout instead of to tmp files.
stdout is captured by nosetests so it will only display if one
asks for it or there are errors in the test.

Tests added for:     changed test for bug19110
make test-suite run? ran nosetests

releases notes: non-user-visible API/code change(s)

Signed-off-by: Ahmon Dancy <dancy@franz.com>