add mixed workload to events test - rfe9493
Added a -m parameter to the events test to run a mixed workload.
Basically this does:

until interrupted:
    while running Phase 4 queries:
        run Phase 3 adds for the next 30 days
        run Phase 5 delete for the oldest 30 days

Tests added for:     events is a test
make test-suite run? no - events test run

<release-note>
rfe9493: added mixed workload to events test

A -m parameter was added to the events test to run a Phase 6
mixed workload after the normal phases complete. This runs
a mixture of queries, adds, and deletes until it is interrupted.
</release-note>

Signed-off-by: Ahmon Dancy <dancy@franz.com>