Python API and examples update
This file incorporates some suggestions and
updates by Ross that were in rfe9432.

<release-note>
Rfe9432 - Updates to Python API document

There were several updates to the Python API documnent
to better delineate the distinction between public and private
methods in 4.0.3.
</release-note>

rfe9540, I have done a method-by-method comparison of the Python API document
with the current Python source files.

<release-note>
rfe9540 - Audited the Python API doc for currency

The Python API document has been updated to reflect recent
changes in the Python API.
</release-note>

bug19210: The examples have been patched to work around the issue with the
limit parameter of getStatements(). I filed a bug and
assigned it to Ross, but I can't find it in BH.

Change-Id: I87b3463d457f18904043779d00adf3841668fb85