rfe9831 - Python tutorial: Size matters.
Alterations to the Python API doc, the Python tutorial,
and tutorial_examples_40.py example10(), to work in
examples of conn.size() reporting on specific subgraphs.

<release-note>
rfe9831 - Python tutorial, illustrate use of conn.size() with contexts.

Example10() of the Python tutorial has been extended to
show examples of the conn.size() method returning the
number of triples found in specific subgraphs.
</release-note>

Change-Id: Ifabcaed32a1f04b1c4b1a42474cff59e95f5d84d
Reviewed-on: https://gerrit.franz.com:9080/662
Reviewed-by: Kevin Layer <layer@franz.com>
Tested-by: Kevin Layer <layer@franz.com>