rfe9802 - Test new dynamic indexing code
As part of the RFE for testing dynamic indexing, the events test
now by default creates only an optimal set of indices to satisfy
the query requirements.

If one wants to use the full set of indices, pass the -i argument
on the command line.

-i, --indexfull       Create the full set of indices [default=optimal
                      indices]

Tests added for:     None
make test-suite run? no - but ran the events test both w and w/o -i

Change-Id: Ic03f5e254cc73a359a809abcd5cd09efb04224ab
Reviewed-on: https://gerrit.franz.com:9080/721
Reviewed-by: Kevin Layer <layer@franz.com>
Tested-by: Kevin Layer <layer@franz.com>