[rfe10093] add profiling to Python events test
Added a -f/--profile option which accepts option strings "off",
"time", or "space" to the python events test. It defaults to "off".

It profiles one of each type of process (1 loader in each loading phase,
1 SPARQL and 1 PROLOG querier, and 1 Deleter), and the flat and call
graph output goes to the agraph.log.

This change requires the rfe10093 change in the agraph repository
as well, which adds (require :prof) to the agraph executable.

Tests added for:     None
make test-suite run? python events test run

Change-Id: I273df2dc941b73ce84ab137d5b4266c231cf992c
Reviewed-on: https://gerrit.franz.com:9080/778
Reviewed-by: Kevin Layer <layer@franz.com>
Tested-by: Kevin Layer <layer@franz.com>