bug19698 - add "SPARQL"->QueryLanguage.SPARQL map prepare*Query
Moved the call to map the strings "SPARQL" and "PROLOG"
to their equivalent QueryLanguage.* instances up a class so the
prepareTupleQuery, prepareBooleanQuery, and prepareGraphQuery
functions all do the conversion.

Tests added for:     changed some QueryLanguage uses to strings
make test-suite run? make python-prepush - yes

<release-note>
bug19698: Changes to prepareGraphQuery and prepareBooleanQuery

prepareGraphQuery and prepareBooleanQuery now also accept the
string "SPARQL" or "PROLOG" as a query language parameter.
Previously, only prepareTupleQuery allowed the string values
as a convenience in place of the QueryLanguage.SPARQL or
QueryLanguage.PROLOG objects.
</release-note>

Change-Id: Ie13b46f9201427d3229dad66ccf9928241cb62fa
Reviewed-on: https://gerrit.franz.com:9080/780
Reviewed-by: John O'Rourke <john.orourke@franz.com>
Reviewed-by: Ahmon Dancy <dancy@franz.com>
Tested-by: Kevin Layer <layer@franz.com>