rfe9428 - Berlin benchmark (with modified whole words test)
Added Python script to run the Berlin Benchmark similarly to:

http://www4.wiwiss.fu-berlin.de/bizer/BerlinSPARQLBenchmark/results/V5/index.html

It an option to run using a full text index instead of regex for
Query 6.

The test does not use or require the Python client, and works with
early versions of Python (2.4) so that there are no installation
requirements on Linux installs other than AllegroGraph.

The rfe says the performance target is same as Virtuoso. We are
roughly 10x faster on queries 7-11, but 10x slower on all others.
I will add a separate RFE for Increasing the server performance
for Berlin queries.

Tests added for:     none - it is a benchmark test
make test-suite run? no

Change-Id: I3f236731dc609f188984b39377e66619bdeb21ac