Support a single load-worker. Remove syncs after query and delete phase.
The delete phase for the events test was designed to delete 10% of the
specified triples using two load-workers, and will hang if a single
worker is created. Use the method employed by the java events test to
allow a single worker to be specified. In this case, only 5% of triples
will be deleted.

Also, remove syncs after the query and delete phases as these are not
needed and brings this test in line with the lisp events test.

make test-suite run? no

releases notes: minor change. no release notes needed.

Change-Id: Ic7e99c46ca1f9bf194faf3c944fd0b202114bb4b
Reviewed-on: https://gerrit.franz.com:9080/1477
Reviewed-by: Ross Jekel <rjekel@franz.com>
Reviewed-by: Ahmon Dancy <dancy@franz.com>
Tested-by: Kevin Layer <layer@franz.com>