rfe11313: remove export*, use response capture
RepositoryConnection.export and exportStatements relied
on client side serialization code for ntriples and
rdf+xml that was buggy and not up to spec. It is not
necessary for us to provide Python client serialization
of triples in several formats, so the functionality has
been replaced with a method to allow the customer to capture
the server's response on an API call in any supported mime
type for that call.

A new context manager function on RepositoryConnection
is called saveResponse, and can be used as follows:

with open('out', 'w') as response:
    with conn.saveResponse(response, 'application/rdf+xml'):
        # The server response is written to response
        conn.getStatements(None, None, None)

export and exportStatements have been rewritten in terms of
this server-side serialization, and have been marked as
deprecated with a warning pointing users to the newer
API.

This will fix the customer's serialization issues
in spr38901.

Tests added for:     rfe11313
make test-suite run? yes - python prepush

<release-note>
rfe11313: remove export*, use response capture

RepositoryConnection.export and exportStatements have been
deprecated. They used client side serialization of statements
that was buggy and not up to spec. The functionality has
been replaced with a way to capture server responses of
any supported mime type.

A new context manager function on RepositoryConnection
is called saveResponse, and can be used as follows:

with open('out', 'w') as response:
    with conn.saveResponse(response, 'application/rdf+xml'):
        # The server response is written to response
        conn.getStatements(None, None, None)

You can reference the function test_save_response
in src2/franz/openrdf/tests/tests.py for examples.
</release-note>

<documentation>
RepositoryConnection.saveResponse(fileobj, accept, raiseAll=False)

Save the server response(s) for the call(s) within the with statement
to fileobj, using accept for the response type requested.

Responses will be uncompressed and saved to fileobj, but not decoded.
Therefore, the API called will likely error out shortly after the
response is saved (which is okay because we really only want
the side-effect of saving the response).

RequestError is always thrown on errors from the server.
Other exceptions can be optionally raised with raiseAll=True.

You will only want to make only one conn call in the with statement
unless you wrap each call in its own try/except.

Example:

with open('out', 'w') as response:
    with conn.saveResponse(response, 'application/rdf+xml'):
        # The response is written to response
        conn.getStatements(None, None, None)

You can reference the function test_save_response
in src2/franz/openrdf/tests/tests.py for examples.
</documentation>

Change-Id: I2bb0ad4a099be58855d728bca4916c748fc9ca99
Reviewed-on: https://gerrit.franz.com:9080/1937
Reviewed-by: Ahmon Dancy <dancy@franz.com>
Tested-by: Kevin Layer <layer@franz.com>