initial python 3 support
tests added    : cloned Python 2 tests for Python 3 tests
tests run      : no
performance    : no change (unless you're using Python 3 in which case
                 things are a bit faster)

<release note>
Support Python 3

AllegroGraph now supports both Python 2 and Python 3. The sources for
Python 2 continue to be in the src2 directory whereas those for Python 3
are in the new src3 directory.
</release note>

Added new directories to hold a port of the Python code from Python 2 to
Python 3. Used 2to3 and much web searching to update the auto-converted
code into something that runs.

Moved stress/bsbm/bsbmtools to bsbmtools so that it can be shared.

Updated (but did not test) stress3/bsbm/bsbm.py and stress3/load/load.
Both of these are quite old and not maintained on the Python 2 side
of things.

Change-Id: If46d4e8efa3d7889f763471c6ee97a39781cda1b
Reviewed-on: https://gerrit.franz.com:9080/5121
Reviewed-by: Gary Warren King <gwking@franz.com>
Reviewed-by: Ahmon Dancy <dancy@franz.com>
Tested-by: Hudson <hudson@franz.com>
Reviewed-by: John O'Rourke <jor@franz.com>