rfe12546: Add getDuplicateStatements to RepositoryConnection
Implements AGRepositoryConnection.getDuplicateStatements in the python
client, which returns the set of duplicate triples in the repository.
Takes a `mode` argument to select how duplicates are
determined. Acceptable values are "spog" or "spo".

Tests added for:     franz/openrdf/tests/tests.py
make test-suite run? python-prepush

<release-note>
rfe12546: Add getDuplicateStatements to AGRepositoryConnection

Adds a new method to the RepositoryConnection class called
getDuplicateStatements, that returns the set of duplicate triples in
the repository. The method signature is:

def getDuplicateStatements(mode)

Duplicate triples are determined by the `mode` argument, which can be
specified as either "spo" (compare as ntriples) or "spog" (compare as
nquads).

</release-note>

<documentation>
see release note.
</documentation>

Change-Id: I376350568aa19718b9304914464a3d6be98353de
Reviewed-on: https://gerrit.franz.com:9080/4173
Reviewed-by: Ahmon Dancy <dancy@franz.com>
Reviewed-by: John O'Rourke <jor@franz.com>
Tested-by: Kevin Layer <layer@franz.com>