bug24290: Python shutdown hang
When a session is started (by way of openSession()) a pingSession()
thread is started to keep the session alive until closeSession() is
called.  If closeSession() is not called by the time Python is ready
to shut down, the pingSession() thread will block the shutdown until
its polling timeout has completed.  This can delay the termination of
Python for a very long time.  This is particularly noticeable when
running the tutorial.

I modified pingSession() so that it marks itself as a daemon thread.
Daemon threads are terminated w/o waiting when Python is ready to shut
down.

<release-note>
bug24290: Python shutdown hang

Previously if a session was created but not closed before Python was
ready to exit, the shutdown of Python could be delayed by a very long
time.  This problem has been fixed.  However, it is always best to
call closeSession() on a session when it is no longer needed.
</release-note>

Tests run: make python-prepush and ran the tutorial.

Change-Id: Idcc5e5a34671bc25749e3ac41e7543eb08050761
Reviewed-on: https://gerrit.franz.com:9080/6605
Tested-by: Hudson <hudson@franz.com>
Reviewed-by: Ahmon Dancy <dancy@franz.com>
Tested-by: Ahmon Dancy <dancy@franz.com>