Simplified version handling in Python
Instead of trying to guess the version from a few different places
the client now uses a hard-coded string in src/franz/__init__.py.

During a release a script checks this number against the AG version
number obtained from the build environment and fails if it detects
discrepancies.

Note that it is OK if the client's number contains something in
addition to the AG version - for example 6.2.0-post1 is a valid
version number when building for AG 6.2.0.

Other changes:
  - Fixed a setup-time warning about the missing VERSION file.
  - Removed windows-support - it was outdated anyway.

Change-Id: Ic15234d62150edbd5f45d3bef1bc6cfb75dd3112
Reviewed-on: https://gerrit.franz.com:9080/6645
Reviewed-by: Tadeusz Sznuk <ts@franz.com>
Tested-by: Hudson <hudson@franz.com>
Reviewed-by: Ahmon Dancy <dancy@franz.com>