rfe14694: Replace cjson with simplejson
Reworked the code to:
   * Try importing simplejson first (might be faster on 2.6)
   * If that is not available use the built-in module

Modified setup.py to allow calls like this:

    pip install agraph-6.2.0-python-client[simplejson]

that will include simplejson in the list of dependencies.

The built-in module should be efficient enough on Python 2.7+,
on Python 2.6 we should recommend simplejson.

Actually, we should recommend not using an ancient and unsupported Python,
but CentOS/RHEL 6 users might not have an alternative.

<release-note>
rfe14694: drop support for cjson

The Python client no longer uses the cjson library. Python 2.6 users
should install the simplejson library instead to avoid a significant
degradation in performance. Other versions of Python are unaffected.
</release-note>

Here's a comparison of the time (in seconds) it takes to retrieve 1000
triples with and without simplejson on Python 2.6:

    N           Min           Max        Median           Avg        Stddev
x 100          0.04          0.06          0.05        0.0488  0.0043298354
+ 100          0.08           0.1          0.09        0.0902  0.0028355606
Difference at 95.0% confidence
        0.0414 +/- 0.00101444
        84.8361% +/- 2.07877%
        (Student's t, pooled s = 0.00365977)

Change-Id: I82037bf7f8f9971effbcbf34a637f60d363709aa
Reviewed-on: https://gerrit.franz.com:9080/6683
Reviewed-by: Tadeusz Sznuk <ts@franz.com>
Tested-by: Hudson <hudson@franz.com>
Reviewed-by: Ahmon Dancy <dancy@franz.com>