rfe14695: Replace pycurl with requests
Note that pycurl is still supported and will be used
if it can be successfully imported. To force the use
of requests even when pycurl is available set the
environment variable AG_FORCE_REQUESTS_BACKEND to
a non-empty value.

We might later drop support for pycurl if performance
tests show that there is no benefit in maintaining it.

Tests are now run two times - once with pycurl and once
with requests.

Also made a few unrelated changes in the readme.

<release-note>
rfe14695 - Replace pycurl with requests.

The HTTP library used by the AllegroGraph Python client
has been changed from pycurl to requests. This change
means that the Python client no longer depends on any
libraries that require native extensions. This results
in a simpler installation process, as it is no longer
necessary to have a working C development environment
on the target machine.

Note that pycurl will still be used if it is installed,
since it might offer better performance in some scenarios.
</release-note>

Change-Id: Ibb7904e825eca47fcd5c91158f8c85a6e904e03d
Reviewed-on: https://gerrit.franz.com:9080/6684
Reviewed-by: Tadeusz Sznuk <ts@franz.com>
Tested-by: Hudson <hudson@franz.com>
Reviewed-by: Ahmon Dancy <dancy@franz.com>