bug24379: HTTPS in Python does not work with Basic auth
Made the url parsing in Server.__init__ more robust
(but still backwards compatible).

Also refactored tests and documentation for ag_connect.

<release-note>
bug24379 - HTTPS in Python does not work with Basic auth

The Python client used to only allow HTTPS connections if
a client certificate was used for authentication.

It is now possible to connect to HTTPS servers by passing
either a `host` value starting with 'https://' or a `protocol`
value of 'https' when creating the server object.
</release-note>

Change-Id: I7c1bd5ab4f674c1070f2375df3083088e44bd454
Reviewed-on: https://gerrit.franz.com:9080/6795
Reviewed-by: Tadeusz Sznuk <ts@franz.com>
Tested-by: Hudson <hudson@franz.com>
Reviewed-by: Ahmon Dancy <dancy@franz.com>
Tested-by: Ahmon Dancy <dancy@franz.com>