bug24322, bug24405: Uploading files + Unicode strings in Python
The Python client used to do file uploads by reading the whole 
file into memory. Unfortunately that is not so easy to fix due 
to how wonderful pycurl is.

The client also failed to ensure that non-ascii text is 
utf-8 encoded before being passed to curl, causing error 
in addData().

<release-note>
bug24322 - Python clients reads imported file into memory

When uploading data to the server the Python client used to
read the whole file into memory before sending it. This 
could cause out of memory errors for larger files. To avoid 
this issue, the client now sends files in chunks. This method 
requires only a limited amount of memory.

bug24405 - Non-ASCII characters in addData()

The Python client used to raise an exception (specifically 
a UnicodeDecodeError) if the data passed to addData was a 
Unicode string containing non-ASCII characters.

This has been corrected.  
</release-note>

Change-Id: I05722c09c412270ab4f5bcd4ded8318d36666951
Reviewed-on: https://gerrit.franz.com:9080/6857
Reviewed-by: Tadeusz Sznuk <ts@franz.com>
Reviewed-by: Ahmon Dancy <dancy@franz.com>
Tested-by: Ahmon Dancy <dancy@franz.com>