rfe14843: Document proxy support in the Python client
<release-note>
rfe14843 - Proxy support in the Python client

It is now possible to configure the Python client to use a proxy when
communicating with the server. This can be done by setting the
AGRAPH_PROXY environment variable to a value of the form
TYPE://HOST:PORT. TYPE can be either 'socks' or 'http'.
</release-note>

Change-Id: Ibe541f2d6ba9d7706f6273a2927fb09ee7b05c38
Reviewed-on: https://gerrit.franz.com:9080/7176
Reviewed-by: Tadeusz Sznuk <ts@franz.com>
Tested-by: Hudson <hudson@franz.com>
Reviewed-by: Ahmon Dancy <dancy@franz.com>