A script for uploading conda packages.
To use the script, you need to:
   - Create an account on anaconda.org
   - Yell at me to add you to the 'franzinc' organization.
   - Just call the script, enter your credentials and
     the sources will be packaged and uploaded to anaconda.org

The script is also invoked during 'make publish'.

Note that his only works in the internal network, since the script
looks for anaconda installations on san1.

Added a proper license classifier to setup.py

Removed a comment from the makefile. The code this comment was
supposed to explain has been removed some time ago.

Change-Id: Ida81ee53bbb743f88e70d0e05e5fb0222f638b5e
Reviewed-on: https://gerrit.franz.com:9080/7332
Reviewed-by: Tadeusz Sznuk <ts@franz.com>
Tested-by: Hudson <hudson@franz.com>
Reviewed-by: Ahmon Dancy <dancy@franz.com>