Do not use git when building a release
`make dist` used to perform all the required version magic
when building a release. Unfortunately git is not available
at that time, so the process was split into 3 steps:

  - Run `make prepare-release`. This will strip .dev from
    the version number, commit and push the result.
    Direct access to the repository (i.e. not through gerrit)
    is required.
  - Make the release in the usual way. `make dist` will check
    if .dev has been stripped from the version number.
  - After the release do `make post-release` to increment
    the version number. The result will again be pushed to git.

If there is any chance of the release making it into the hands
of a customer please make sure you do `make publish` before
the third step (or you can just checkout the correct revision
and do `make publish` later).

Changed VERSION to AGVERSION in the makefile.

Added a release-process.md document.

Change-Id: I5c1921ba47515f2736e98363a878f64238b284fd
Reviewed-on: https://gerrit.franz.com:9080/7430
Reviewed-by: Tadeusz Sznuk <ts@franz.com>
Tested-by: Hudson <hudson@franz.com>
Reviewed-by: Ahmon Dancy <dancy@franz.com>