bug24443: getStatements sometimes breaks for raw Python values
<release-note>
bug24443 - getStatements sometimes breaks for raw Python values

The getStatements function can be called with a raw Python value,
instead of an RDF term, as its third argument (object filter).
For instance, the following call:

    conn.getStatements(None, somePredicate, True)

would retrieve all triples with the predicate of somePredicate
and the object equal to "true"^^xsd:boolean.

This mechanism did not work correctly if the second argument
(predicate filter) was None. In that case an AttributeError
exception would be raised.

This has been corrected.
</release-note>

Change-Id: I8329ccf60804ce0e1fc1972940457215e967ec93
Reviewed-on: https://gerrit.franz.com:9080/7439
Reviewed-by: Tadeusz Sznuk <ts@franz.com>
Tested-by: Hudson <hudson@franz.com>
Reviewed-by: Ahmon Dancy <dancy@franz.com>