bug24892: Fix time parsing
It's time to fix this.

The problem is that the code assumes - and explicitly asserts - that
times returned by AllegroGraph will always be in UTC. When that is
not the case an AssertionError is raised.

Simply removing the assertion would not fix things, since Python's
standard library lacks any implementation of timezone objects that
would be required to represent non-UTC times. So I've replaced our
custom datetime parsing with an external library (that does provide
a timezone implementation).

Change-Id: Ifac9be5b618c057b93d9a9129f9e54337b152cc1
Reviewed-on: https://gerrit.franz.com:9080/8444
Reviewed-by: Tadeusz Sznuk <ts@franz.com>
Tested-by: Hudson <hudson@franz.com>
Reviewed-by: Hudson <hudson@franz.com>
Reviewed-by: Ahmon Dancy <dancy@franz.com>