More robust virtualenv update
we used to pass the --recreate flag to tox when we detected that
the .venv file is older than one of the requirements files.

This doesn't work in the following scenario:
   1. requirements are modified.
   2. a target that uses some, but not all, of the environments is used.
      An example would be prepush2 or prepush3
   3. A different target is invoked later. The .venv file has been
      refreshed already, so tox will not install new dependencies.
   4. People yell at me on Slack.

To avoid this issue the makefile now completely deletes ALL
environments whenever the requirements change.

Also updated one requirement, so getting this change will help people
who already went through step 2.

Change-Id: I53a37f12139335b2461eb12a72561f4518d45f3a
Reviewed-on: https://gerrit.franz.com:9080/8507
Reviewed-by: Tadeusz Sznuk <ts@franz.com>
Tested-by: Hudson <hudson@franz.com>
Reviewed-by: Hudson <hudson@franz.com>
Reviewed-by: Ahmon Dancy <dancy@franz.com>