Do not use the system-wide python interpreter
Since it tends to be too ancient on some machines.
Also we want to test on all versions, not just those that happen
to be available at a particular testing machine.

Instead we install anaconda and use that to create environments for
each version of Python we want to test on. All installation steps
are performed offline if SAN1 is available - otherwise packages
and installers are downloaded from the internet.

Also added a new 'make clean' target.

Change-Id: Ie47ede53f0a21dd809e55883011acdad0d6f09ea
Reviewed-on: https://gerrit.franz.com:9080/9298
Reviewed-by: Tadeusz Sznuk <ts@franz.com>
Tested-by: Hudson <hudson@franz.com>
Reviewed-by: Hudson <hudson@franz.com>
Reviewed-by: Ahmon Dancy <dancy@franz.com>