Jupyter-friendly stdout
On Python 3 you have to use `sys.stdout.buffer` instead of
`sys.stdout` to do binary I/O. The output_to context manager does this
automatically when you pass True as the argument. But in IDLE and
Jupyter sys.stdout is a different object (used by these tools to
capture the output) that does not have `.buffer` - but seems to accept
binary data.  So I've changed output_to to check if stdout has the
`.buffer` attribute instead of just checking if we're on Python 3.

Change-Id: I78c2dc509302b384cb16df7834cbecf6982810be
Reviewed-on: https://gerrit.franz.com:9080/9779
Reviewed-by: Tadeusz Sznuk <ts@franz.com>
Tested-by: Hudson <hudson@franz.com>
Reviewed-by: Hudson <hudson@franz.com>
Reviewed-by: Ahmon Dancy <dancy@franz.com>