Distributed transaction settings
Distributed transaction settings can now be configured
for a connection. It is also possible to provide settings
for a single commit (as parameters to ``commit()``) or
a code block (using a new context manager).

The version number has been bumped to 100.1.0.

Note that I have used camelCase while naming the new methods
to maintain consistency with the old code, but a bit of my
soul dies every time I have to type such a method name.

Change-Id: Ie37234fd87550294a18a1faa6ee7bea7fdc9ed4f
Reviewed-on: https://gerrit.franz.com:9080/10371
Reviewed-by: Tadeusz Sznuk <ts@franz.com>
Tested-by: Hudson <hudson@franz.com>
Reviewed-by: Hudson <hudson@franz.com>
Reviewed-by: Ahmon Dancy <dancy@franz.com>