Fix when passing proxy configuration as argument to AllegroGraphServer or agconnect.
1. We should create an attribute in 'session' called 'proxies' instead of 'proxy'
since the class site-packages/requests.Session() would be expecting 'proxies'
as an optional argument and has triggers for it but not for 'proxy'.
We process the scheme the same way it was done and put it in 'proxy_server'
but now we will assign it to self.proxies in a dictionary:
{'http' : proxy_server} since the only protocol we use is 'http'.

2. We can still use AGRAPH_PROXY the same way after the fix.

3. We can bypass this by setting the environment variable HTTP_PROXY with the URL
of the proxy we would like to use. This way we will make the package requests to use the
proxy we have set in HTTP_PROXY.
But! this way we would make all of the traffic that uses the requests package
to access the internet through it.

Iván D. Ramos V.