bug24876: RowReader issues
Resolved said issues by removing that buggy class.
It was only used by a single method, and the interface of
that method is so weak and poorly documented it is
unlikely that anyone has ever used it, so no changelog
entry for this fix.

We still need a working solution for incremental parsing.
Unfortunately all I could find was between 2x (for
solutions that require C-based components) and 40x
(pure Python) slower than non-incremental parsing.

Change-Id: Ib9cccc177564008165a42e4b34f1865dc7bd8add
Reviewed-on: https://gerrit.franz.com:9080/11451
Reviewed-by: Tadeusz Sznuk <ts@franz.com>
Tested-by: Hudson <hudson@franz.com>
Reviewed-by: Ahmon Dancy <dancy@franz.com>