Ensure sessions are closed properly
Sessions created via the ``openFederation()`` or ``openSession()``
methods of franz.openrdf.sail.allegrographserver.AllegroGraphServer
were not closed when ``close()`` was called on the resulting
connection.  This has been fixed.

Updated CHANGES.rst with notes for this change and the dependencies
updated due to the github-reported security vuln's.

Change-Id: I75c7ad6cd86ea7f6cea3aea6aadd78380e7499e6
Reviewed-on: https://gerrit.franz.com:9080/12992
Reviewed-by: Mikel Bancroft <mikel@franz.com>
Tested-by: Hudson <hudson@franz.com>
PTEST: Hudson <hudson@franz.com>
Reviewed-by: Ahmon Dancy <dancy@franz.com>