Corrected test problems in example019.rst of python tutorial
If you have TestCode and TestOutput labels over code sections,
then the code bits must be executable and the results match the
output bits. This rule was not followed in the example019.rst
file and as a result the file failed to pass the verification tests.
Change so the code sections are indicated by :: (prepended by a space)
at the end of the paragraph before the code. Now no code is compared
to output. The rendered file looks the same.

Change-Id: Iac97eaeef33caf064ba5f5f2afbdd68f2bd21d62
Reviewed-on: https://gerrit.franz.com:9080/13126
Reviewed-by: David Margolies <dm@franz.com>
Reviewed-by: Ahmon Dancy <dancy@franz.com>
Tested-by: Ahmon Dancy <dancy@franz.com>