Update python tests to be compatible with gerrit:13924
gerrit:13924 changes the format of strings passed in the indexLiterals
arg when creating a freetext index. This change updates the python
test suite so that it passes with this change in place.

Updates the docstrings for modifyFreeTextIndex and createFreeTextIndex
to reflect his change.

Tests added for:       no
make test-suite run?   yes
Documentation changes? no

releases notes: non-user-visible API/code change(s)

Change-Id: Ib6eec6a0d89637b40475b6863af6ffcf09ae3f42
Reviewed-on: https://gerrit.franz.com:9080/13932
Reviewed-by: Mikel Bancroft <mikel@franz.com>
Reviewed-by: Ahmon Dancy <dancy@franz.com>
Tested-by: Ahmon Dancy <dancy@franz.com>