bug25881: update .travis.yml to fix failing builds
It appears that due to Travis CI infrastructure changes the /dev/shm
hack is 1) no longer working and 2) not needed, so this commit removes
it. It also updates to the matrix to include AG 6.5.0 and 6.6.0 and
Python 3.5, 3.6 and 3.7 releases.

Tests added for:       no
make test-suite run?   no
Documentation changes? no

Change-Id: I33bd020c4c755f74bea54cccd714b2374ebd121d
Reviewed-on: https://gerrit.franz.com:9080/14114
Reviewed-by: Yuri Zhykin <yuriz@franz.com>
Reviewed-by: Ahmon Dancy <dancy@franz.com>
Tested-by: Ahmon Dancy <dancy@franz.com>