Modified examples that depended on the order of query results
A few examples depended on the order of query results from the server
though the order is undefined in most cases.  I remedied this by
adding ORDER BY clauses to queries as needed.  For the
enableDuplicateFilter() example I just instructed Sphinx to not verify
the test output.

Change-Id: I6b7dc2194c17f46ce38e98cd6404d82bff9dd1c6
Reviewed-on: https://gerrit.franz.com:9080/c/agraph-python/+/15740
Tested-by: Hudson <hudson@franz.com>
PTEST: Hudson <hudson@franz.com>
Reviewed-by: Ahmon Dancy <dancy@franz.com>