rfe16555: Add support for handling 204 responses from AG
Properly handle 204 No Content responses from AG by not trying to read
or transform (e.g. as JSON) the response body, or by treating it as an
empty JSON object.

This is to support a related change in AG in which a 204 response is
sent for empty triple-result and row-result requests.

No tests added. Already covered by testBasics in
src/franz/miniclient/test.py

Change-Id: I5976e94211c2cf4e4e27b9bb9155f9c345dbc819
Reviewed-on: https://gerrit.franz.com:9080/c/agraph-python/+/17364
Reviewed-by: Mikel Bancroft <mikel@franz.com>
Reviewed-by: John Foderaro <jkf@franz.com>
Tested-by: Hudson <hudson@franz.com>
PTEST: Hudson <hudson@franz.com>