AG-666: Improve agraph-python project (dependencies) management
Switching to use pyproject.toml for project management. Most of the
metadata of setuptools and pytest are migrated into pyproject.toml so
these files can be removed:

+ pytest.ini
+ requirements.txt
+ setup.cfg
+ setup.py

As a result, to build the wheel as well as a source distribution, one
needs to run `python -m build`; to install all dependencies for
development, one needs to run (hopefully in a virtualenv environment)
`pip install -e ".[dev]"`.

Also, support for Python 3.6 has been removed in this change.

Change-Id: I9f1ea87fe079b93ea97182cf2e1988471b2e11cf
Reviewed-on: https://gerrit.franz.com:9080/c/agraph-python/+/24678
Tested-by: Hudson <hudson@franz.com>
Ptest: Hudson <hudson@franz.com>
Reviewed-by: Tianyu Gu <gty@franz.com>
Reviewed-by: John Foderaro <jkf@franz.com>
Reviewed-by: Yuri Zhykin <yuriz@franz.com>