AG-165: eliminate Python 2 related legacy code
This changeset comprehensively eliminates legacy Python 2 code. As a
result, `future` and `six` are removed from the dependencies.

Note that it also affects source code that are under `docs/` folder. As
an side effect, `black` and `isort` had been enabled and caused an
one-off modifications.

Change-Id: Icdb341be7f63c2d67f21317b356bd46d15b1a2ff
Reviewed-on: https://gerrit.franz.com:9080/c/agraph-python/+/24710
Reviewed-by: Yuri Zhykin <yuriz@franz.com>
Reviewed-by: John Foderaro <jkf@franz.com>
Reviewed-by: Tianyu Gu <gty@franz.com>
Tested-by: Hudson <hudson@franz.com>
Ptest: Hudson <hudson@franz.com>